# doc/swagger.json
{
    "swagger": "2.0",
    "basePath": "/v2",
    "paths": {
...
        "/collections/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "summary": "List all collections of kind document from DB",
                "operationId": "get_collections",
                "tags": [
                    "collections"
                ]
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Create collection of kind document in DB",
                "operationId": "post_collections",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Collections"
                        }
                    }
                ],
                "tags": [
                    "collections"
                ]
            }
        },
        "/collections/search/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Search document in collections of kind document from DB",
                "operationId": "get_search",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Items number per page",
                        "default": 10
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "type": "string",
                        "description": "Query",
                        "default": "[[{\"field\":\"name\",\"operator\":\"LIKE\",\"value\":\"\"}]]"
                    },
                    {
                        "name": "collections",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Collections"
                    }
                ],
                "tags": [
                    "collections"
                ]
            }
        },
        "/collections/{collection}/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                },
                {
                    "description": "Name Of Collection",
                    "name": "collection",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Document Already Exists"
                    }
                },
                "summary": "Search documents from collection",
                "operationId": "get_collection",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Items number per page",
                        "default": 10
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "type": "string",
                        "description": "Query",
                        "default": "[[{\"field\":\"name\",\"operator\":\"LIKE\",\"value\":\"\"}]]"
                    }
                ],
                "tags": [
                    "collections"
                ]
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Document Already Exists"
                    }
                },
                "summary": "Insert document in DB",
                "operationId": "post_collection",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/DocumentPost"
                        }
                    }
                ],
                "tags": [
                    "collections"
                ]
            }
        },
        "/collections/{collection}/clear/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                },
                {
                    "description": "Name Of Collection",
                    "name": "collection",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Clear documents in collection",
                "operationId": "post_collection_clear",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/DocumentClear"
                        }
                    }
                ],
                "tags": [
                    "collections"
                ]
            }
        },
        "/collections/{collection}/{key}/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                },
                {
                    "description": "Name Of Collection",
                    "name": "collection",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "description": "Key Of Document",
                    "name": "key",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Delete document by key",
                "operationId": "delete_document",
                "tags": [
                    "collections"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Get document by key",
                "operationId": "get_document",
                "tags": [
                    "collections"
                ]
            },
            "patch": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Partial update document",
                "operationId": "patch_document",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/DocumentPatch"
                        }
                    }
                ],
                "tags": [
                    "collections"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Update document",
                "operationId": "put_document",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/DocumentPut"
                        }
                    }
                ],
                "tags": [
                    "collections"
                ]
            }
        },
        "/edges/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "summary": "List all collections of kind edge from DB",
                "operationId": "get_edges",
                "tags": [
                    "edges"
                ]
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Create collection of kind edge in DB",
                "operationId": "post_edges",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Edge"
                        }
                    }
                ],
                "tags": [
                    "edges"
                ]
            }
        },
        "/edges/search/": {
            "parameters": [
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Search edge in collections of kind edge from DB",
                "operationId": "get_edge_search",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Items number per page",
                        "default": 10
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "type": "string",
                        "description": "Query",
                        "default": "[[{\"field\":\"name\",\"operator\":\"LIKE\",\"value\":\"\"}]]"
                    },
                    {
                        "name": "edges",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Edges Document"
                    }
                ],
                "tags": [
                    "edges"
                ]
            }
        },
...
        "/edges/{edge}/clear/": {
            "parameters": [
                {
                    "description": "Name Of Edge(Collection)",
                    "name": "edge",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "required": true,
                    "default": "Token token=",
                    "in": "header",
                    "description": "Token Authorization",
                    "name": "Authorization",
                    "type": "string"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "summary": "Clear documents in edge",
                "operationId": "post_edge_clear",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/EdgeClear"
                        }
                    }
                ],
                "tags": [
                    "edges"
                ]
            }
        },
...
    },
    "info": {
        "title": "GloboMap API",
        "version": "2.0",
        "description": "GloboMap API"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
...
    "definitions": {
...
        "Collections": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
                "name": {
                    "type": "string",
                    "pattern": "^([a-z_]){1,}$",
                    "minimum": 1
                }
            },
            "required": [
                "name"
            ]
        },
        "DocumentPost": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
                "id": {
                    "type": "string",
                    "pattern": "^([a-zA-Z0-9=._-]){1,}$",
                    "minimum": 1
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string",
                    "pattern": "^([a-z_]){1,}$",
                    "minimum": 1
                },
                "timestamp": {
                    "type": "number"
                },
                "properties": {
                    "type": "object"
                },
                "properties_metadata": {
                    "type": "object"
                }
            },
            "required": [
                "id",
                "provider",
                "timestamp"
            ]
        },
        "DocumentClear": {
            "type": "array",
            "items": {
                "type": "array",
                "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                        "field": {
                            "type": "string",
                            "minimum": 1
                        },
                        "operator": {
                            "type": "string",
                            "enum": [
                                "LIKE",
                                "IN",
                                "==",
                                "!=",
                                ">",
                                ">=",
                                "<",
                                "<="
                            ]
                        },
                        "value": {
                            "type": [
                                "string",
                                "boolean",
                                "integer",
                                "array",
                                "object"
                            ],
                            "minimum": 1
                        }
                    },
                    "required": [
                        "field",
                        "operator",
                        "value"
                    ]
                }
            }
        },
        "DocumentPatch": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
                "id": {
                    "type": "string",
                    "pattern": "^([a-zA-Z0-9=._-]){1,}$",
                    "minimum": 1
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string",
                    "pattern": "^([a-z_]){1,}$",
                    "minimum": 1
                },
                "timestamp": {
                    "type": "number"
                },
                "properties": {
                    "type": "object"
                },
                "properties_metadata": {
                    "type": "object"
                }
            }
        },
...
    },
...
}